fixed another deadlock issue
print was called in enter callback in main thread. enter callback was called in key callback which lock mutex at the start so when print is called in the enter callback a deadlock is formed. so to fix it unlock the mutex before calling the enter callback and relock it after enter callback is done.